requirements.txt
duckdb>=1.4.4
fastapi>=0.116.0
langchain>=0.3.27
langchain-openai>=0.3.28
pydantic>=2.11.0
python-dotenv>=1.1.0
uvicorn>=0.35.0
